rename the publicPosts folder fr publiPosts to publicPoosts
